+++ include/tools/codegen.h
@@ -5,7 +5,8 @@
 #include <string>
 #include <type_traits>
 
-#include "tools/fnv.h"
+#include "fnv.h"
+#include "string_helpers.h"
 
 namespace codegen {
     constexpr char kTabSym = '\t';
@@ -25,6 +26,13 @@ namespace codegen {
         {128, "uint128_t"},
         {256, "uint256_t"},
         {512, "uint512_t"},
+    };
+    // clang-format on
+
+    // clang-format off
+    inline std::initializer_list<std::pair<std::string, std::string>> kTypeNameReplaceRules = {
+        {"< ", "<"},
+        {" >", ">"},
     };
     // clang-format on
 
@@ -204,18 +212,11 @@ namespace codegen {
             return push_line(std::format("// {}", text), move_cursor_to_next_line);
         }
 
-        static void replace_all(std::string& text, const std::string& to_remove, const std::string& to_insert) {
-            auto pos = text.find(to_remove);
-            while (pos != std::string::npos) {
-                text.replace(pos, to_remove.length(), to_insert);
-                pos = text.find(to_remove, pos);
-            }
-        }
-
         self_ref prop(const std::string& type_name, const std::string& name, bool move_cursor_to_next_line = true) {
             auto type_name_fixed = type_name;
-            replace_all(type_name_fixed, "< ", "<");
-            replace_all(type_name_fixed, " >", ">");
+            for (auto [search, replace] : kTypeNameReplaceRules) {
+                replace_all(type_name_fixed, search, replace);
+            }
             const auto line = move_cursor_to_next_line ? std::format("{} {};", type_name_fixed, name) : std::format("{} {}; ", type_name_fixed, name);
             return push_line(line, move_cursor_to_next_line);
         }
